requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

asgiref==3.8.1
django==5.0.7
djangorestframework==3.15.2
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1; sys_platform == "win32"
